WIP: Refactor + `:core` README
