+++ core/data-model/src/main/java/com/ivy/data/Settings.kt
@@ -3,6 +3,7 @@ package com.ivy.data
 import java.math.BigDecimal
 import java.util.*
 
+@Deprecated("won't use")
 data class Settings(
     val theme: Theme,
     val baseCurrency: CurrencyCode,
