+++ core/data-model/src/main/java/com/ivy/data/Value.kt
@@ -1,6 +1,4 @@
-package com.ivy.data.transaction
-
-import com.ivy.data.CurrencyCode
+package com.ivy.data
 
 data class Value(
     val amount: Double,
