+++ core/data-model/src/main/java/com/ivy/data/transaction/OverdueSection.kt
@@ -1,5 +1,7 @@
 package com.ivy.data.transaction
 
+import com.ivy.data.Value
+
 data class OverdueSection(
     val income: Value,
     val expense: Value,
