+++ core/data-model/src/main/java/com/ivy/data/transaction/Transaction.kt
@@ -1,6 +1,7 @@
 package com.ivy.data.transaction
 
 import com.ivy.data.SyncState
+import com.ivy.data.Value
 import com.ivy.data.account.Account
 import com.ivy.data.attachment.Attachment
 import com.ivy.data.category.Category
