+++ core/data-model/src/main/java/com/ivy/data/transaction/TrnListItem.kt
@@ -1,5 +1,6 @@
 package com.ivy.data.transaction
 
+import com.ivy.data.Value
 import java.time.LocalDate
 
 sealed interface TrnListItem {
