+++ core/domain/src/main/java/com/ivy/core/domain/pure/dummy/DummyDateDivider.kt
@@ -1,8 +1,8 @@
 package com.ivy.core.functions.transaction
 
 import com.ivy.common.dateNowUTC
+import com.ivy.data.Value
 import com.ivy.data.transaction.TrnListItem
-import com.ivy.data.transaction.Value
 import java.time.LocalDate
 
 fun dummyDateDivider(
