+++ core/ui/src/main/java/com/ivy/core/ui/value/ValueFunctions.kt
@@ -2,7 +2,7 @@ package com.ivy.core.ui.value
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import com.ivy.data.transaction.Value
+import com.ivy.data.Value
 import com.ivy.wallet.utils.format
 import com.ivy.wallet.utils.shortenAmount
 import com.ivy.wallet.utils.shouldShortAmount
